requirements.txt
black>=24.10.0
scikit-learn>=1.6.0
numpy>=2.2.1
matplotlib>=3.10.0
tensorflow>=2.18.0
spacy>=3.8.3
wordcloud>=1.9.4
textblob>=0.18.0.post0
SpeechRecognition>=3.12.0
opencv-python>=4.10.0.84
pytesseract>=0.3.13
pillow>=11.0.0